Refactor code style for user profile training
